refactor: move Observer interfaces to entities layer
