+++ src/entities/state/index.ts
@@ -2,3 +2,5 @@ export { default as State } from "./model/StatePublisher";
 
 // TODO: get rid of this state altogether, and use StatePublisher instead
 export { default as StateAuth } from "./model/State";
+
+export { type Publisher, type Observer } from "./model/Observer";
