+++ src/entities/state/model/Observer.ts
