+++ src/features/game/controls/RoundControls.ts
@@ -1,6 +1,6 @@
 import { Component, label, option, select } from "../../../shared/ui";
 import { isValidSetting } from "../../../shared/helpers";
-import { Observer, Publisher } from "../../../shared/Observer";
+import { Observer, Publisher } from "../../../entities/state";
 
 import LevelsState from "../model/LevelsState";
