+++ src/features/game/fields/GameField.ts
@@ -4,7 +4,7 @@ import AssembleRow from "./AssembleRow";
 import RoundState from "../model/RoundState";
 import HintSettings from "../model/HintSettings";
 
-import { Observer, Publisher } from "../../../shared/Observer";
+import { Observer, Publisher } from "../../../entities/state";
 import {
   ANIMATION_DELAY_COEFFICIENT,
   calculateImageAspectRatio,
