+++ src/features/game/fields/Row.ts
@@ -1,6 +1,6 @@
 import { Component, div } from "../../../shared/ui";
 import { debounceListener } from "../../../shared/helpers";
-import { Observer, Publisher } from "../../../shared/Observer";
+import { Observer, Publisher } from "../../../entities/state";
 
 import WordCard, { type Word } from "../card/WordCard";
