+++ src/features/game/hints/HintControls.ts
@@ -1,7 +1,7 @@
 import { Component, i, input, label } from "../../../shared/ui";
 import HintSettings, { HintSettingsData } from "../model/HintSettings";
 
-import { Observer, Publisher } from "../../../shared/Observer";
+import { Observer, Publisher } from "../../../entities/state";
 import { isValidSetting } from "../../../shared/helpers";
 
 import styles from "./HintControls.module.css";
