+++ src/features/game/hints/TranslationHint.ts
@@ -3,7 +3,7 @@ import { Component } from "../../../shared/ui";
 import RoundState from "../model/RoundState";
 import HintSettings from "../model/HintSettings";
 
-import { Observer, Publisher } from "../../../shared/Observer";
+import { Observer, Publisher } from "../../../entities/state";
 
 import styles from "./TranslationHint.module.css";
